% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{dtt_add_units}
\alias{dtt_add_units}
\alias{dtt_add_years}
\alias{dtt_add_months}
\alias{dtt_add_days}
\alias{dtt_add_hours}
\alias{dtt_add_minutes}
\alias{dtt_add_seconds}
\title{Add Units}
\usage{
dtt_add_units(x, units, n = 1L)

dtt_add_years(x, n = 1L, ...)

dtt_add_months(x, n = 1L, ...)

dtt_add_days(x, n = 1L, ...)

dtt_add_hours(x, n = 1L, ...)

dtt_add_minutes(x, n = 1L, ...)

dtt_add_seconds(x, n = 1L, ...)
}
\arguments{
\item{x}{A date time vector.}

\item{units}{A string of the units.}

\item{n}{An integer of the number of units.}

\item{...}{Unused.}
}
\value{
The modified date time vector.
}
\description{
Add time units to a date time vector.
}
\examples{
dtt_add_units(as.Date("1999-12-31"), "days")
}
\seealso{
\code{\link[=dtt_subtract_units]{dtt_subtract_units()}}
}
