% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{dtt_adjust_units}
\alias{dtt_adjust_units}
\title{Adjust Units}
\usage{
dtt_adjust_units(x, from = "seconds", to = "seconds")
}
\arguments{
\item{x}{An integer or numeric vector}

\item{from}{A string of the original units.}

\item{to}{A string of the new units.}
}
\value{
A numeric vector.
}
\description{
Adjust Units
}
\examples{
dtt_adjust_units(60, to = "minutes")
}
\seealso{
\code{\link[=dtt_add_units]{dtt_add_units()}}

\code{\link[=dtt_subtract_units]{dtt_subtract_units()}}

Other units: 
\code{\link{dtt_units_per_unit}()},
\code{\link{dtt_units}()}
}
\concept{units}
