% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completed.R
\name{dtt_completed}
\alias{dtt_completed}
\alias{dtt_completed.Date}
\alias{dtt_completed.POSIXct}
\alias{dtt_completed.hms}
\title{Completed}
\usage{
dtt_completed(x, ...)

\method{dtt_completed}{Date}(x, units = "days", unique = TRUE, sorted = TRUE, ...)

\method{dtt_completed}{POSIXct}(x, units = "seconds", unique = TRUE, sorted = TRUE, ...)

\method{dtt_completed}{hms}(x, units = "seconds", unique = TRUE, sorted = TRUE, ...)
}
\arguments{
\item{x}{A date/time vector.}

\item{...}{Unused.}

\item{units}{A string of the time units. The possible values are "secs",
"minutes", "hours", "days" or "weeks".}

\item{unique}{A flag indicating whether the values must be unique.}

\item{sorted}{A flag indicating whether the values must be sorted.}
}
\value{
A flag indicating whether complete.
}
\description{
Tests whether a date time is complete.
}
\section{Methods (by class)}{
\itemize{
\item \code{dtt_completed(Date)}: Test if Date vector is complete

\item \code{dtt_completed(POSIXct)}: Test if POSIXct vector is complete

\item \code{dtt_completed(hms)}: Test if POSIXct vector is complete

}}
\seealso{
Other complete: 
\code{\link{dtt_complete}()}
}
\concept{complete}
