% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-add-time.R
\name{dtt_date_add_time}
\alias{dtt_date_add_time}
\title{Date Add Time}
\usage{
dtt_date_add_time(x, time, tz = dtt_default_tz())
}
\arguments{
\item{x}{A Date vector.}

\item{time}{A hms vector.}

\item{tz}{A string of the timezone.}
}
\value{
A POSIXct vector.
}
\description{
Adds times to Dates vector and sets timezone in a single function.
}
\examples{
dtt_date_add_time(as.Date("2001-03-05"), hms::as_hms("06:07:08"), tz = "Etc/GMT+9")
}
