% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-from-ints.R
\name{dtt_date_from_ints}
\alias{dtt_date_from_ints}
\title{Create dates from integers}
\usage{
dtt_date_from_ints(year = 1970L, month = 1L, day = 1L)
}
\arguments{
\item{year}{An integer of the year. The default value is 1970.}

\item{month}{An integer of the month between 1 and 12. The default value is
the 1st month.}

\item{day}{An integer of the day between 1 and 31. The default value is the
1st day.}
}
\value{
A floored Date vector.
}
\description{
Create dates from integers
}
\examples{
dtt_date_from_ints(
 year = 1991,
 month = 07,
 day = 23
)

dtt_date_from_ints(
 year = c(1991, 1992, 1993),
 month = c(07, 07, 07),
 day = c(23, 24, 21)
)
}
