% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day-decimal.R
\name{dtt_day_decimal}
\alias{dtt_day_decimal}
\title{Get Decimal Day Values}
\usage{
dtt_day_decimal(x, ...)
}
\arguments{
\item{x}{A date/time vector.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Gets decimal day values for date/time vectors.
}
\examples{
x <- as.POSIXct("1990-01-03 10:00:01")
dtt_day_decimal(x)

}
