% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days-in-month.R
\name{dtt_days_in_month}
\alias{dtt_days_in_month}
\title{Days in the Month}
\usage{
dtt_days_in_month(x)
}
\arguments{
\item{x}{A Date or POSIXct vector.}
}
\value{
A integer vector of 28 to 31 indicating the days in the month.
}
\description{
Days in the Month
}
\examples{
dtt_days_in_month(as.Date(c("2000-02-11", "2001-02-01")))
}
\seealso{
Other days: 
\code{\link{dtt_days_in_year}()},
\code{\link{dtt_doy_to_date}()},
\code{\link{dtt_doy}()}
}
\concept{days}
