% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days-in-year.R
\name{dtt_days_in_year}
\alias{dtt_days_in_year}
\title{Days in the Year}
\usage{
dtt_days_in_year(x)
}
\arguments{
\item{x}{A Date or POSIXct vector.}
}
\value{
A integer vector of 365 and 366 indicates the days of the year.
}
\description{
Days in the Year
}
\examples{
dtt_days_in_year(as.Date(c("2000-10-11", "2001-01-01")))
}
\seealso{
Other days: 
\code{\link{dtt_days_in_month}()},
\code{\link{dtt_doy}()},
\code{\link{dtt_doy_to_date}()}
}
\concept{days}
