% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days-of-year.R
\name{dtt_days_of_year}
\alias{dtt_days_of_year}
\title{Days of the Year}
\usage{
dtt_days_of_year(x)
}
\arguments{
\item{x}{A Date or POSIXct vector.}
}
\value{
A integer vector of 365 and 366 indicates the days of the year.
}
\description{
Days of the Year
}
\examples{
dtt_days_of_year(as.Date(c("2000-10-11", "2001-01-01")))
}
