% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayte-time.R
\name{dtt_dayte_time}
\alias{dtt_dayte_time}
\alias{dtt_dayte_time.POSIXct}
\title{Dayte Time}
\usage{
dtt_dayte_time(x, ...)

\method{dtt_dayte_time}{POSIXct}(x, start_month = 1L, start_day = 1L,
  ...)
}
\arguments{
\item{x}{A date time vector.}

\item{...}{Unused.}

\item{start_month}{An integer vector specifying the start month of the year.}

\item{start_day}{An integer vector specifying the start day of the year.}
}
\value{
A Date vector with the year set to year.

A Date vector of the daytes.
}
\description{
Dayte Time
}
\section{Methods (by class)}{
\itemize{
\item \code{POSIXct}: Dayte Time a POSIXct vector
}}

\examples{
dtt_dayte_time(as.POSIXct(c("2001-01-01 12:13:14", "2015-12-13"), tz = "Etc/GMT+10"))
}
