% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{dtt_daytt}
\alias{dtt_daytt}
\title{Dayte Time}
\usage{
dtt_daytt(x, start_month = 1L, start_day = 1L)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{start_month}{An integer vector specifying the start month of the year.}

\item{start_day}{An integer vector specifying the start day of the year.}

\item{...}{Unused.}
}
\value{
A Date or POSIXct vector with the year for February 29th as 1972.
}
\description{
Dayte Time
}
