% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decade.R
\name{dtt_decade}
\alias{dtt_decade}
\alias{dtt_decade.Date}
\title{Decade}
\usage{
dtt_decade(x, ...)

\method{dtt_decade}{Date}(x, ...)
}
\arguments{
\item{x}{A date/time vector.}

\item{...}{Unused.}
}
\value{
A integer vector of the decade.
}
\description{
Decade
}
\section{Methods (by class)}{
\itemize{
\item \code{Date}: Decade a Date vector
}}

\examples{
dtt_decade(as.Date(c("2001-01-01", "2015-12-13")))
}
