% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/tz.R
\name{dtt_sys_tz}
\alias{dtt_sys_tz}
\alias{dtt_set_sys_tz}
\alias{dtt_reset_sys_tz}
\alias{dtt_default_tz}
\alias{dtt_set_default_tz}
\alias{dtt_reset_default_tz}
\title{Get, Set or Reset Default Time Zone}
\usage{
dtt_sys_tz()

dtt_set_sys_tz(tz = NULL)

dtt_reset_sys_tz()

dtt_default_tz()

dtt_set_default_tz(tz = NULL)

dtt_reset_default_tz()
}
\arguments{
\item{tz}{A string of the time zone.}
}
\value{
A string of the current or old time zone.
}
\description{
Get, Set or Reset Default Time Zone
}
\section{Functions}{
\itemize{
\item \code{dtt_set_default_tz()}: Set Default Time Zone

\item \code{dtt_reset_default_tz()}: Reset Default Time Zone

}}
\examples{
\dontrun{
dtt_default_tz()
old <- dtt_set_default_tz("Etc/GMT+8")
dtt_default_tz()
dtt_reset_default_tz()
dtt_default_tz()
dtt_set_default_tz(old)
dtt_default_tz()
}
}
\seealso{
Other tz: 
\code{\link{dtt_adjust_tz}()},
\code{\link{dtt_set_tz}()},
\code{\link{dtt_tz}()}
}
\concept{tz}
