% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy-decimal.R
\name{dtt_doy_decimal}
\alias{dtt_doy_decimal}
\title{Day of the Year Decimal}
\usage{
dtt_doy_decimal(x, ...)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{...}{Unused.}
}
\value{
A numeric vector between 0 and 366 of the day of the year.
}
\description{
Day of the Year Decimal
}
\examples{
dtt_doy_decimal(Sys.Date())
}
\seealso{
\code{\link[=dtt_doy]{dtt_doy()}}

Other decimal: 
\code{\link{dtt_day_decimal}()},
\code{\link{dtt_hour_decimal}()},
\code{\link{dtt_minute_decimal}()},
\code{\link{dtt_month_decimal}()},
\code{\link{dtt_year_decimal}()}
}
\concept{decimal}
