% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy.R
\name{dtt_doy_to_date}
\alias{dtt_doy_to_date}
\title{Day of the Year to Date}
\usage{
dtt_doy_to_date(x, year = 1972L)
}
\arguments{
\item{x}{An integer vector of the Day of the Year.}

\item{year}{An integer scalar or vector of the year.}
}
\value{
A Date vector.
}
\description{
Day of the Year to Date
}
\examples{
dtt_doy_to_date(3L)
}
