% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_to_time.R
\name{dtt_excel_to_time}
\alias{dtt_excel_to_time}
\title{Convert Excel times to times.}
\usage{
dtt_excel_to_time(x, ...)
}
\arguments{
\item{x}{A vector of numbers to convert.}

\item{...}{Unused.}
}
\value{
A floored hms vector.
}
\description{
Converts Excel times encoded as a fraction of a day to hms class.
}
\examples{
dtt_excel_to_time(0.5)
dtt_excel_to_time(0.25)
dtt_excel_to_time(0.75)
}
\seealso{
Other floor: 
\code{\link{dtt_date}()},
\code{\link{dtt_date_time}()},
\code{\link{dtt_excel_to_date}()},
\code{\link{dtt_floor}()},
\code{\link{dtt_floored}()},
\code{\link{dtt_set_time}()}
}
\concept{floor}
