% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feb29-to-28.R
\name{dtt_feb29_to_28}
\alias{dtt_feb29_to_28}
\title{Convert Feb 29 to Feb 28}
\usage{
dtt_feb29_to_28(x)
}
\arguments{
\item{x}{A Date or POSIXct vector.}
}
\value{
The modified Date or POSIXct vector.
}
\description{
Converts Feb 29 to Feb 28th
}
\examples{
dtt_feb29_to_28(as.Date("2004-02-29"))
}
\seealso{
Other leap year: 
\code{\link{dtt_leap_year}()}
}
\concept{leap year}
