% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{dtt_is_date_time}
\alias{dtt_is_date_time}
\title{Is Date Time}
\usage{
dtt_is_date_time(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A flag indicating whether R is a POSIXct vector.
}
\description{
Is Date Time
}
\seealso{
Other is: 
\code{\link{dtt_is_date}()},
\code{\link{is_date_time}()}
}
\concept{is}
