% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{dtt_is_dtt}
\alias{dtt_is_dtt}
\title{Is Date or DateTime Object}
\usage{
dtt_is_dtt(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A flag indicating whether R is a Date or POSIXct vector.
}
\description{
Is Date or DateTime Object
}
