% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/month.R, R/set.R
\name{dtt_months}
\alias{dtt_months}
\alias{dtt_months<-}
\alias{dtt_month}
\alias{dtt_month<-}
\alias{dtt_month.Date}
\alias{dtt_month.POSIXct}
\alias{dtt_month<-.Date}
\alias{dtt_month<-.POSIXct}
\alias{dtt_set_month}
\title{Get and Set Month Values}
\usage{
dtt_months(x, ...)

dtt_months(x) <- value

dtt_month(x, ...)

dtt_month(x) <- value

\method{dtt_month}{Date}(x, ...)

\method{dtt_month}{POSIXct}(x, ...)

\method{dtt_month}{Date}(x) <- value

\method{dtt_month}{POSIXct}(x) <- value

dtt_set_month(x, value)
}
\arguments{
\item{x}{A date/time vector.}

\item{...}{Unused.}

\item{value}{A integer vector of the month value(s).}
}
\value{
An integer vector (or the modified date/time vector).
}
\description{
Gets and sets month values for date/time vectors.
}
\section{Methods (by class)}{
\itemize{
\item \code{Date}: Get integer vector of month values for a Date vector

\item \code{POSIXct}: Get integer vector of month values for a POSIXct vector

\item \code{Date}: Set month values for a Date vector

\item \code{POSIXct}: Set month values for a POSIXct vector
}}

\examples{
x <- as.Date("1990-01-02")
dtt_month(x)
dtt_month(x) <- 11L
x

x <- as.POSIXct("1990-01-02 23:40:51")
dtt_month(x)
dtt_month(x) <- 11L
x
}
