% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month-decimal.R
\name{dtt_month_decimal}
\alias{dtt_month_decimal}
\title{Get Decimal Month Values}
\usage{
dtt_month_decimal(x, ...)
}
\arguments{
\item{x}{A date/time vector.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Gets decimal month values for date/time vectors.
}
\examples{
x <- as.POSIXct("1990-01-03 10:00:01")
dtt_month_decimal(x)

}
