% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-year.R
\name{dtt_study_year}
\alias{dtt_study_year}
\title{Study Year}
\usage{
dtt_study_year(x, start = 1L)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{start}{An integer scalar of the starting month or a Date scalar of the starting dayte.}
}
\value{
A character vector of the study year.
}
\description{
Study Year
}
\examples{
dtt_study_year(as.Date(c("2000-03-31", "2000-04-01", "2001-04-01")), start = 4L)
}
