% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.R
\name{dtt_subtract_units}
\alias{dtt_subtract_units}
\alias{dtt_subtract_years}
\alias{dtt_subtract_months}
\alias{dtt_subtract_days}
\alias{dtt_subtract_hours}
\alias{dtt_subtract_minutes}
\alias{dtt_subtract_seconds}
\title{Subtract Time Units}
\usage{
dtt_subtract_units(x, n = 1L, units = dtt_units(x))

dtt_subtract_years(x, n = 1L)

dtt_subtract_months(x, n = 1L)

dtt_subtract_days(x, n = 1L)

dtt_subtract_hours(x, n = 1L)

dtt_subtract_minutes(x, n = 1L)

dtt_subtract_seconds(x, n = 1L)
}
\arguments{
\item{x}{A date/time vector.}

\item{n}{An integer of the number of units.}

\item{units}{A string of the time units. The possible values are "secs",
"minutes", "hours", "days" or "weeks".}
}
\value{
The modified date time vector.
}
\description{
Subtract time units from a date time vector.
}
\examples{
dtt_subtract_units(as.Date("1999-12-31"), 2L, "days")
}
\seealso{
\code{\link[=dtt_add_units]{dtt_add_units()}}
}
\concept{subtract}
