% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-date.R
\name{dtt_sys_date}
\alias{dtt_sys_date}
\title{Get System Date}
\usage{
dtt_sys_date(tz = dtt_default_tz())
}
\arguments{
\item{tz}{A string of the time zone.}
}
\value{
A floored Date scalar.
}
\description{
Get System Date
}
\examples{
\dontrun{
dtt_set_default_tz("Etc/GMT+12")
dtt_sys_date()
dtt_set_default_tz("Etc/GMT-12")
dtt_sys_date()
dtt_sys_date(tz = "Etc/GMT+12")
}
}
\seealso{
Other sys: 
\code{\link{dtt_sys_date_time}()},
\code{\link{dtt_sys_time}()}
}
\concept{sys}
