% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-date-time.R
\name{dtt_sys_date_time}
\alias{dtt_sys_date_time}
\title{Get System Date Time}
\usage{
dtt_sys_date_time(tz = dtt_default_tz())
}
\arguments{
\item{tz}{A string of the time zone.}
}
\value{
A floored POSIXct scalar.
}
\description{
Get System Date Time
}
\examples{
\dontrun{
dtt_set_default_tz("UTC")
dtt_sys_date_time()
dtt_set_default_tz("Etc/GMT+8")
dtt_sys_date_time()
dtt_sys_date_time(tz = "UTC")
}
}
\seealso{
Other sys: 
\code{\link{dtt_sys_date}()},
\code{\link{dtt_sys_time}()}
}
\concept{sys}
