% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys-time.R
\name{dtt_sys_time}
\alias{dtt_sys_time}
\title{Get System Time}
\usage{
dtt_sys_time(tz = dtt_default_tz())
}
\arguments{
\item{tz}{A string of the time zone.}
}
\value{
A floored hms scalar.
}
\description{
Get System Time
}
\examples{
\dontrun{
dtt_sys_time()
}
}
\seealso{
Other sys: 
\code{\link{dtt_sys_date_time}()},
\code{\link{dtt_sys_date}()}
}
\concept{sys}
