% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{dtt_timer}
\alias{dtt_timer}
\alias{dtt_start}
\alias{dtt_stop}
\alias{dtt_elapsed}
\alias{dtt_reset}
\alias{dtt_is_running}
\title{Use a dtt_timer}
\usage{
dtt_timer(seconds = 0L, start = FALSE)

dtt_start(x)

dtt_stop(x)

dtt_elapsed(x)

dtt_reset(x, seconds = 0L, start = FALSE)

dtt_is_running(x)
}
\arguments{
\item{seconds}{A count of the start time in seconds.}

\item{start}{A flag indicating whether to start the timer.}

\item{x}{A hms scalar.}
}
\description{
Use a dtt_timer
}
\section{Functions}{
\itemize{
\item \code{dtt_start}: Start Timer

\item \code{dtt_stop}: Stop Timer

\item \code{dtt_elapsed}: Elapsed Time

\item \code{dtt_reset}: Reset Timer

\item \code{dtt_is_running}: Is Running
}}

\examples{
# create a new timer
t <- dtt_timer()

# no time has elapsed because the timer has not started
dtt_elapsed(t)

# start the timer
t <- dtt_start(t)

# get the time elapsed
# time elapsed is increasing because the timer is still running
Sys.sleep(1)
dtt_elapsed(t)
Sys.sleep(1)
dtt_elapsed(t)

# stop the timer
t <- dtt_stop(t)

# time elapsed is now fixed
dtt_elapsed(t)
Sys.sleep(1)
dtt_elapsed(t)

# reset the timer
t <- dtt_reset(t)
dtt_elapsed(t)
}
