% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz.R
\name{dtt_tz}
\alias{dtt_tz}
\alias{dtt_tz.POSIXct}
\title{Get, Set or Adjust Time Zone}
\usage{
dtt_tz(x, ...)

\method{dtt_tz}{POSIXct}(x, ...)
}
\arguments{
\item{x}{A date/time vector.}

\item{...}{Unused.}
}
\value{
A string of the time zone.
}
\description{
Gets, sets or  the time zone for a date time vector.
}
\section{Methods (by class)}{
\itemize{
\item \code{POSIXct}: Get the time zone for a POSIXct vector.
}}

\examples{
dtt_tz(as.POSIXct("1970-01-01", tz = "Etc/GMT+8"))
}
\seealso{
Other tz: 
\code{\link{dtt_adjust_tz}()},
\code{\link{dtt_set_tz}()},
\code{\link{dtt_sys_tz}()}
}
\concept{tz}
