% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{dtt_units_per_unit}
\alias{dtt_units_per_unit}
\title{Units per Unit}
\usage{
dtt_units_per_unit(units = "seconds", unit = "days")
}
\arguments{
\item{units}{A string of the time units. The possible values are "secs",
"minutes", "hours", "days" or "weeks".}

\item{unit}{A string of the time unit.}
}
\value{
A number of the units per unit
}
\description{
Units per Unit
}
\examples{
dtt_units_per_unit("hours")
}
\seealso{
Other units: 
\code{\link{dtt_adjust_units}()},
\code{\link{dtt_units}()}
}
\concept{units}
