% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year-decimal.R
\name{dtt_year_decimal}
\alias{dtt_year_decimal}
\title{Get Decimal Year Values}
\usage{
dtt_year_decimal(x, ...)
}
\arguments{
\item{x}{A date/time vector.}

\item{...}{Unused.}
}
\value{
A numeric vector.
}
\description{
Gets decimal year values for date/time vectors.
}
\examples{
x <- as.Date("1990-01-02")
dtt_year_decimal(x)

}
