% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dttr-package.R
\docType{package}
\name{dttr2-package}
\alias{dttr2}
\alias{dttr2-package}
\title{dttr2: Dates, Times and DateTimes}
\description{
Simple manipulations and conversions for date (Date), date time (POSIXct) 
    and time (hms) objects.
    As well as functions to get, set, add and subtract time units it includes 
    functions to floor objects, set or adjust time zones and complete sequences. 
    It also provides a timer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/dttr}
  \item Report bugs at \url{https://github.com/poissonconsulting/dttr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

}
\keyword{internal}
