% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.datetime}
\alias{is.datetime}
\alias{is.POSIXct}
\alias{is.Date}
\alias{is.hms}
\title{Is Date/Time}
\usage{
is.POSIXct(x)

is.Date(x)

is.hms(x)
}
\arguments{
\item{x}{An object}
}
\value{
A flag indicating whether x inherits from Date, POSIXct or hms.
}
\description{
Tests whether an object is a Date, POSIXct, or hms vector.
}
