% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_date_time}
\alias{is_date_time}
\alias{is.POSIXct}
\alias{is.Date}
\alias{is_date}
\alias{is.hms}
\alias{is_time}
\title{Is Date/Time}
\usage{
is.POSIXct(x)

is_date_time(x)

is.Date(x)

is_date(x)

is.hms(x)

is_time(x)
}
\arguments{
\item{x}{An object}
}
\value{
A flag indicating whether x inherits from Date, POSIXct or hms.
}
\description{
Tests whether an object is a Date, POSIXct, or hms vector.
}
\seealso{
Other is: 
\code{\link{dtt_is_date}()},
\code{\link{dtt_is_date_time}()}
}
\concept{is}
