% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\docType{data}
\name{NA_hms_}
\alias{NA_hms_}
\title{Missing hms}
\format{An object of class \code{hms} (inherits from \code{difftime}) of length 1.}
\usage{
NA_hms_
}
\description{
A missing hms object
}
\keyword{datasets}
