% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\docType{data}
\name{NA_POSIXct_}
\alias{NA_POSIXct_}
\title{Missing POSIXct}
\format{
An object of class \code{POSIXct} (inherits from \code{POSIXt}) of length 1.
}
\usage{
NA_POSIXct_
}
\description{
A missing POSIXct object
}
\keyword{datasets}
