% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericise.R
\name{numericise.hms}
\alias{numericise.hms}
\title{Numericise (or Numericize)}
\usage{
\method{numericise}{hms}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A double numeric atomic vector object.
}
\description{
Coerce an R object to a numeric atomic object.
}
\details{
Unlike the other functions in the package
for handling date time objects, \code{numericise()} does not floor times.
}
\examples{
numericise(hms::as_hms("00:01:03"))
}
