% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{params}
\alias{params}
\title{Parameters for dttr2 functions}
\arguments{
\item{units}{A string of the time units. The possible values are
"secs", "minutes", "hours", "days" or "weeks".}
}
\description{
Descriptions of the parameters for the dttr2 functions.
}
\keyword{internal}
