% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vld.R
\name{vld_time}
\alias{vld_time}
\title{Validate Time}
\usage{
vld_time(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A flag indicating whether the condition was met.
}
\description{
Validates that an object is scalar \link[hms:hms]{hms::hms} object
using
\code{inherits(x, class) && length(x) == 1L && !anyNA(x)}.
}
\examples{
vld_time(1)
vld_time(hms::as_hms("10:12:59"))
}
\seealso{
\code{\link[=chk_time]{chk_time()}}
}
