% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_time}
\alias{chk_time}
\title{Check Time}
\usage{
chk_time(x, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
\code{NULL}, invisibly. Called for the side effect of throwing an error
if the condition is not met.
}
\description{
Checks if scalar hms object using \code{\link[=vld_time]{vld_time()}}.
}
\examples{
chk_time(hms::as_hms("10:00:10"))
try(chk_time(1))
}
