% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayte.R
\name{dtt_dayte}
\alias{dtt_dayte}
\alias{dtt_dayte.Date}
\alias{dtt_dayte.POSIXct}
\title{Dayte}
\usage{
dtt_dayte(x, ...)

\method{dtt_dayte}{Date}(x, start = 1L, ...)

\method{dtt_dayte}{POSIXct}(x, start = 1L, ...)
}
\arguments{
\item{x}{A date/time vector.}

\item{...}{Unused.}

\item{start}{An integer scalar of the starting month or a Date scalar of the starting date.}
}
\value{
A Date vector with the year set to year.

A Date vector of the daytes.
}
\description{
Dayte
}
\section{Methods (by class)}{
\itemize{
\item \code{Date}: Dayte a Date vector

\item \code{POSIXct}: Dayte a POSIXct vector
}}

\examples{
dtt_dayte(as.Date(c("2001-01-01", "2015-12-13")))
}
\seealso{
Other dayte: 
\code{\link{dtt_dayte_time}()},
\code{\link{dtt_daytt}()}
}
\concept{dayte}
