% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{dtt_daytt}
\alias{dtt_daytt}
\title{Dayte Time}
\usage{
dtt_daytt(x, start = 1L)
}
\arguments{
\item{x}{A Date or POSIXct vector.}

\item{start}{An integer vector of the starting month or a Date vector of the
starting date.}
}
\value{
A Date or POSIXct vector with the year for February 29th as 1972.
}
\description{
Dayte Time
}
\seealso{
Other dayte: 
\code{\link{dtt_dayte_time}()},
\code{\link{dtt_dayte}()}
}
\concept{dayte}
