% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{dtt_units_per_unit}
\alias{dtt_units_per_unit}
\title{Units per Unit}
\usage{
dtt_units_per_unit(units = "seconds", unit = "days")
}
\arguments{
\item{units}{A string of the time units.}

\item{unit}{A string of the time unit.}
}
\value{
A number of the units per unit
}
\description{
Units per Unit
}
\examples{
dtt_units_per_unit("hours")
}
