% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dttr-package.R, R/dttr2-package.R
\docType{package}
\name{dttr2-package}
\alias{dttr2}
\alias{dttr2-package}
\title{dttr2: Manipulate Dates, DateTimes and Times}
\description{
Manipulates date (Date), datetime (POSIXct) and time (hms) vectors.
    Date/times are considered discrete and are floored whenever encountered.
    Times are wrapped and time zones are maintained unless explicitly altered 
    by the user.

Manipulates date (Date), datetime (POSIXct) and time (hms) vectors.
    Date/times are considered discrete and are floored whenever encountered.
    Times are wrapped and time zones are maintained unless explicitly altered 
    by the user.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/dttr}
  \item Report bugs at \url{https://github.com/poissonconsulting/dttr/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/dttr}
  \item Report bugs at \url{https://github.com/poissonconsulting/dttr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})


\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

}
\keyword{internal}
