% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dttr-package.R, R/dttr2-package.R
\docType{package}
\name{dttr2-package}
\alias{dttr2}
\alias{dttr2-package}
\title{dttr2: Manipulate Date, POSIXct and hms Vectors}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Manipulates date ('Date'), date time ('POSIXct') and time
    ('hms') vectors.  Date/times are considered discrete and are floored
    whenever encountered.  Times are wrapped and time zones are maintained
    unless explicitly altered by the user.

\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Manipulates date ('Date'), date time ('POSIXct') and time
    ('hms') vectors.  Date/times are considered discrete and are floored
    whenever encountered.  Times are wrapped and time zones are maintained
    unless explicitly altered by the user.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/dttr2}
  \item Report bugs at \url{https://github.com/poissonconsulting/dttr2/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/dttr2}
  \item Report bugs at \url{https://github.com/poissonconsulting/dttr2/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}


\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
